created new files for columns
