+++ ramp-up/src/Utilities/ValidateUser.ts
@@ -1,11 +1,30 @@
 import { GridRowModel } from "@mui/x-data-grid";
 import parsePhoneNumber from 'libphonenumber-js';
 
-
-
 const validateUser = (user: GridRowModel, requiredFields: string[]) => {
-    console.log(requiredFields)
-    
+    console.log('validateUser', user, requiredFields)
+
+    if (requiredFields.includes('name') ) {
+        console.log('name', user.name, )
+        return false;
+    }
+    if (requiredFields.includes('age') && !validateAge(user.age)) {
+        console.log('age', user.age, )
+        return false;
+    }
+    if (requiredFields.includes('mobile') && !validateMobile(user.mobile)) {
+        console.log('mobile', user.mobile, )
+        return false;
+    }
+    if (requiredFields.includes('address') && !validateAddress(user.address)) {
+        console.log('address', user.address, )
+        return false;
+    }
+    if (requiredFields.includes('birthday') && !validateBirthday(user.birthday)) {
+        console.log('birthday', user.birthday, )
+        return false;
+    }
+    console.log('validatedUser', user, requiredFields)   
     return true
 };
 
@@ -36,6 +55,9 @@ const validateMobile = (inputValue: string) => {
       const phoneNumberObj = parsePhoneNumber(inputValue);
       return (phoneNumberObj!.isValid());
     } catch (error) {
+      if (inputValue.startsWith('0') && inputValue.length === 10 && /^\d+$/.test(inputValue)) {
+        return true;
+      }
       return false;
     }
   };
@@ -46,6 +68,7 @@ const validateBirthday = (birthday:Date) => {
     }
     return true;
 }
+
